Loading page to game page routing
